trigger handle_DOWN callback when a new leader is elected if the previous one is down.
Since the handle_DOWN callback is only called in the leader, when the leader goes down, that callback is never called.
So the way to solve this is to keep track of the previous leader so that when a new one is elected, the handke_DOWN callback can be called if the previous leader is still down.