Make some imported lists-functions explicit.
The three methods from lists: keydelete/3, keysearch/3 and foreach/2
are only referenced once in the file. By making them explicit, we
simplify the readability of the code. Furthermore, this change makes the
style in the file consistent. Several places prefixed calls explicitly
with the lists module, while others did not.

Change member/2 and foldl/3 from lists to explicit naming.

This marks the complete removal of all -import() statements in the
code.