Do not use a _ catch all wildcard for lists:member/2.
The lists:member/2 function will return either true or false. Hence,
if it does not return false, it MUST be true that is returned. It is
bad style to catch-all this in general when the enumeration is
bounded. Further extensions to the enumeration will, rather than fail,
return something. Which can be something unwanted.